+++ src/index.ts
@@ -22,7 +22,7 @@ export * from './utils/hooks';
 export * from './utils/timeAndDate';
 export * from './utils/utils';
 
-export * from './login/Login';
+export * from './components/login/LoginPage';
 
 export * from './localization/localization';
 export * from './localization/LocalizationContext';